+++ src/gen_cranberry/gc_dev_batt.cpp
@@ -4,13 +4,13 @@ void gc_dev_batt_open(void){
     pinMode(_PIN_GC_BATT_, INPUT);
 }
 
-int gc_dev_batt_read_raw(void){
-  int val = analogRead(_PIN_GC_BATT_);
+uint16_t gc_dev_batt_read_raw(void){
+  uint16_t val = analogRead(_PIN_GC_BATT_);
   return val;
 }
 
-int gc_dev_batt_read(void){
+uint16_t gc_dev_batt_read(void){
     float raw = (float)analogRead(_PIN_GC_BATT_) * (5.0/1023.0);
-    int val = raw * 1000;
+    uint16_t val = raw * 1000;
     return val;
 }
